+++ semcore/radio/src/Radio.jsx
@@ -5,6 +5,7 @@ import assignProps, { callAllEventHandlers } from '@semcore/utils/lib/assignProp
 import keyboardFocusEnhance from '@semcore/utils/lib/enhances/keyboardFocusEnhance';
 import resolveColorEnhance from '@semcore/utils/lib/enhances/resolveColorEnhance';
 import getInputProps, { inputProps } from '@semcore/utils/lib/inputProps';
+import { useColorResolver } from '@semcore/utils/lib/use/useColorResolver';
 
 import style from './style/radio.shadow.css';
 import logger from '@semcore/utils/lib/logger';
@@ -264,15 +265,18 @@ RadioMark.displayName = 'RadioMark';
 
 const Text = (props) => {
   const SText = Root;
-  const { styles } = props;
+  const { styles, color } = props;
 
   React.useEffect(() => {
     if (props.rootDisabled !== props.disabled) {
       props.hoistDisabled(props.disabled);
     }
   }, [props.rootDisabled, props.disabled, props.hoistDisabled]);
+  const resolveColor = useColorResolver();
 
-  return sstyled(styles)(<SText render={TypographyText} tag='span' />);
+  return sstyled(styles)(
+    <SText render={TypographyText} tag='span' use:color={resolveColor(color)} />,
+  );
 };
 Text.displayName = 'Text';